client-python/requirements.txt
python-engineio
python-socketio
six
websocket-client
requests
aioconsole
aiohttp